+ & -, extract helper functions
