+++ pages/components/Calculator/Calculator.jsx
@@ -1,35 +1,12 @@
 import { useMemo, useCallback, useState } from 'react';
 import _ from 'lodash';
-import math from 'mathjs';
 import styles from './calculator.module.scss';
+import { calculate, findIndexOfLastOperand, findIndexOfLastNumber, trimLastOperand } from './functions';
 
 export default function Calculator() {
   const [display, setDisplay] = useState("0");
   const [equation, setEquation] = useState("0");
   const [isNewNumber, setIsNewNumber] = useState(true);
-
-  const findIndexOfLastOperand = (eq) => {
-    // return the index of the last ops in the input equation
-    // -1 is returned if there's no ops
-    let i;
-    for (i = eq.length-1; i <= 0; i--) {
-      if (['/', 'x', '-', '+'].indexOf(eq[i]) >= 0 && eq[i+1] === " ") break;
-    }
-    return i;
-  }
-
-  const trimLastOperand = (eq) => {
-    if (findIndexOfLastOperand === eq.length-2) eq = eq.substring(0, eq.length-3);
-    return eq;
-  }
-
-  const findIndexOfLastNumber = (eq) => {
-    // return the starting index of the last number in the input equation
-    const lastOpsIndex = findIndexOfLastOperand(trimLastOperand(eq));
-    if (lastOpsIndex < 0) return 0;
-    // if (lastOpsIndex === eq.length-2) return findIndexOfLastNumber(eq.substring(0, eq.length-3));
-    return lastOpsIndex + 2;
-  }
 
   const toggleNegativity = () => {
     // Analyse the equation and add the "-" sign to the appropriate place
@@ -45,8 +22,6 @@ export default function Calculator() {
     // else setEquation(result);
     // setDisplay(result);
   }
-
-  const calculate = (eq) => math.evaluate(trimLastOperand(eq).replace(" ", "").replace("x", "*"));
 
   const handleNumber = (num) => {
     if (isNewNumber || equation[equation.length-2] === "=") {
@@ -83,10 +58,16 @@ export default function Calculator() {
         break;
       
       case "-":
-        
+        eq = trimLastOperand(eq);
+        if (findIndexOfLastOperand(eq) !== -1) setDisplay(calculate(eq));
+        setEquation(eq + " - ");
+        setIsNewNumber(false);
         break;
       case "+":
-        
+        eq = trimLastOperand(eq);
+        if (findIndexOfLastOperand(eq) !== -1) setDisplay(calculate(eq));
+        setEquation(eq + " + ");
+        setIsNewNumber(false);
         break;
   
       case "C":
